Arel::Nodes::DistinctFrom
