+++ lib/arel/extensions/distinct_from.rb
@@ -1,0 +1,22 @@
+# rubocop:disable Naming/MethodName
+# rubocop:disable Naming/UncommunicativeMethodParamName
+
+module Arel
+  module Nodes
+    class DistinctFrom < Arel::Nodes::Binary
+    end
+  end
+
+  module Visitors
+    class ToSql
+      def visit_Arel_Nodes_DistinctFrom(o, collector)
+        visit o.left, collector
+        collector << ' IS DISTINCT FROM '
+        visit o.right, collector
+      end
+    end
+  end
+end
+
+# rubocop:enable Naming/MethodName
+# rubocop:enable Naming/UncommunicativeMethodParamName
